@@ -1,3 +1,3 @@
 module ylang
 
-go 1.17
+go 1.19
